2 - Wrap in an enclosing trait and instantiate here
First wrap the `UserRepository` in an enclosing trait and instantiate
the user repository there. This creates a namespace for our repository.
Next, in order to declare that we would like to have the
`userRepository` instance injected in the `UserService` we will first do
what was done to the other repository; wrap it in an enclosing
(namespace) trait and use a "self-type annotation" to declare our need
for the `UserRepository` service.